Update navbar component to update selected item
Update navbar component to use a callback to update the selected item.
Previously no mechanism to update the selected item was implemented.
The callback uses the dcc.Location component to be notified whenever
navigation occurs and updates the navbar item list based on that.

To simplify the implementation of the callback a change was made in
navbar item generation. Instead of selecting selected item by name it is
now selected by path.