Update navbar item generation tests
After implementing navbar callbacks, some design considerations were
brought up.

Most importantly, highlighting is now done by url, and highlighting a
url that is not in the navbar is supposed to not highlight any elements.
Previously the expected behaviour was to return an empty list.

Additionally, page registry items that should be included in the navbar
are now are expected to be marked with `nav_item=True`.

This commit adds tests for above mentioned behaviours, and improves
other tests.