Fix navbar height and page scrolling
Fix navbar height and page scrolling by making the outermost div the
same height as the screen, and setting the page container to
`overflow-auto`.

This fixes navbar component's height, which previously was not the same
as the screen. Setting the page container to `overflow-auto` makes the
navbar stay in place while scrolling on long pages.