Fix navbar width bug
Fix bug where navbar had different width depending on the selected item.
The navbar's width was set by the widest child element. When that
element was selected it's width was increased by 4px since a border is
added to the selected navbar item.

This commit fixes that issue by adding a margin of equal width as the
border to elements that are not selected.