Fix misleading icon argument
The size argument to the icon component did not change the font size,
but the optical size, which was not obvious and was not stated in
documentation. This commit removes that parameter, instead setting
optical sizing by automatic css rule.

Also adds example on how to *actually* set size for icons.