Fix loading bug in icon component
Due to icons using ligature fonts, the icon component had incorrect
width during loading, leading to unexpected behaviour.

To fix this, setting icon size also sets max width. To accomplish this
in a simple manner, the interface was updated so that size is passed as
a parameter instead of as a class name.