Can draw a graph from multiple csv files
Can take multiple csv files, convert them in polars dataframes and store
them in dcc.store. These files can then be read in main and iterated over.
A figure is then created for each dataframe and stored in a list. This
list is used to create a go.Figure object with a layout.
A dcc.graph object is then created with the go.Figure object.

Need to style download buttons, user naming input and choosing color.